+++ docs/fr_FR/Changelog.rst
@@ -3,6 +3,7 @@ ChangeLog
 
 - Placement réseau: changements en cours.
 - Scripts: Correction droits 'installPackage.sh'
+- Interne: Abeille: setPIGpio().
 
 240223-BETA-1
 -------------